samples: matter: revert DTS overlays for Wi-Fi coex on nRF7002 EK
When a user builds a Matter sample with SHIELD=nrf7002_ek,
the SHIELD variable is not automatically propagated to the
hci_rpmsg child image and hence the nrf_radio_coex device
tree node is not defined in the child image. To workaround
that, I added appropriate DTS overlays for hci_rpmsg child
images. However, this duplicates the definition of the coex
node and hence puts extra maintenance burden.

Revert the DTS overlays.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>