.github: validate manifest userdata
Add a script for checking the userdata fields, using a new pykwalify
schema. Make sure that the semantic contents of the userdata fields
are sensible -- at the very least, we need 'upstream-sha' to be an
ancestor of the SHA pointed to by 'revision', or we've gone stale for
sure.

This ensures that the userdata which were previously added to west.yml
will not bit-rot.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>