scripts: west ncs-compare: use userdata when comparing projects
Update the ncs-compare implementation to consider projects with
upstreams in the zephyrproject-rtos github organization separately
from projects with other upstreams specified via userdata in
nrf/west.yml.

This makes the output more accurate in some cases, such as openthread.
It also makes the job easier when upmerging these projects manually,
and aids in decision making on when that is necessary (because now we
can update upstream-sha in nrf/west.yml and use the ncs-compare
results to decide what we need to do).

Adding an --all switch forces the command to include output even for
projects which say 'compare-by-default: false' in their userdata.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>