applications: serial_lte_modem: Fix use of uninitialized identity_type
When the param_count value is less than or equal to 4, the
identity_type parameter remained uninitialized. This was later passed
on to lwm2m_carrier_identity_write which may lead to unknown behaviour.
This is now fixed. The function do_carrier_portfolio now returns failure
if the param_count value was less than or equal to 4 for READ and WRITE
operations.

Found by sonarcloud as a violation of MISRA C:2004, 9.1

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>