samples: wifi: Fix dropping of scan results
With the new network management queueing design, queuing of events and
running callbacks will happen in parallel, so, if the callbacks take
more time (in this case shell_printf) then queue can overflow.

In earlier design the queue add will be blocked on completion of
callbacks as they used to share the same lock.

So, for now increase the queue size to maximum scan results for shell
sample. This might not be a problem with other samples/applications that
don't print scan results to UART, so, only doing this for select
samples.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>