tfm_platform_system.c: don't include zephyr header
This is breaking the build since zephyr commit
fbf851cdc4d8f0afe60131aea6154945f126d778
("util.h: Add dependency on time_units.h"), which pulled
in an additional Zephyr header that contains inline function
definitions that use various Zephyr-specific Kconfig symbols
that are not available when we're building this file.

The include doesn't seem to be needed since removing it
entirely restores the build, so I'm assuming it's just cruft
that's safe to delete.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>