bluetooth: fast_pair: storage: improve manager section registration
Added a const specifier to the STRUCT_SECTION_ITERABLE macro that is
used to register section entries for the Fast Pair Factory Reset
module. The const specifier should be added for sections that are
placed in ROM according to the API documentation.

Added also a static specifier to the section entry registration macro.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>