net: lib: lwm2m_client_utils: Location assistance result
Add a callback for the result code. Application can register this
callback if the automatic resending handler has been initialized to
get notified about the location assistance results. If the automatic
resending handler hasn't been initialized, the callback will not be
called. If application doesn't initialize the resending handler, the
application is responsible for the resending logic and reading the
corresponding result codes from the objects itself.

Signed-off-by: Jarno Lämsä <jarno.lamsa@nordicsemi.no>