samples: net: mqtt: Add modem initialization
Modem initialization at boot time was deprecated, so the application
was started without modem initialized, causing a reboot loop.

Fixed by adding modem initialization in the network thread.
In addition, readout of IMEI and credential provisioning had
to be moved to be certain that it happens after modem is on.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>