samples: nrf9160: modem_shell: Increase workqueue stack size
Increase MOSH_COMMON_WORKQUEUE_STACK_SIZE to 9126 to accommodate
commands sent from cloud.
Normal commands are run in shell thread but commands sent from cloud
are run in modem shell's own common thread.
This means we need to increase the stack for the common thread a lot
just because of the commands sent from the cloud as we cannot run
them in the same thread as normal shell commands.

Jira: MOSH-444

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>