Bluetooth: controller: Add enable of BT_CTLR_DF_SUPPORT
The host DF related features depend on BT_CTLR_DF_SUPPORT.
That feature was not selected when build of DF samples in the past
because it was dependent on BT_LL_SW_SPLIT and enabled not supported
by SDC features. That in total caused a build errors due to missing
Host APIs (not compiled in due to BT_DF being disabled).

Problematic KConfigs were changed in Zephyr, so now the BT_CTLR_DF-
_SUPPORTED is not dependent on BT_LL_SW_SPLIT. Now SDC can select it
and enable BT_DF to compile in required Host features.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>