Bluetooth: Controller: Add BT_LL_SOFTDEVICE_HEADERS_INCLUDE Kconfig
Add new Kconfig option that enables include of SoftDevice header
files delivered with library.

The SoftDevice header files provided in nrfxlib are included in two
situations: when there is a BT_LL_SOFTDEVICE_BUILD_TYPE_LIB selected
or when !BT_CTRL (Host only build). In case of build from sources
these header files should not be included, that works out of the box
for compound builds (Host and Controller in single binary). In case
of split builds, it should be possible to select which header files
are used by Host.

Mind that this is a dependency that is not verified by build system.
One should remember to select the same header files for APP and NET
cores in case of split builds.

Remove add of SOFTDEVICE_CONTROLLER_LIB_HEADERS to Zephyr link
libraries. It should be added only in case the softdevice_controller
is enables, hence it is moved to nrfxlib/softdevice_controller/-
-CMakeLists. If the library is added here, CONFIG_BT is enabled
but SDC is disabled, build fails due to missing SOFTDEVICE_CONTROLLER-
_LIB_HEADERS.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>