clang-format: Sync clang-format file with Zephyr
cp ~/ncs/zephyr/.clang-format ~/ncs/nrf/.clang-format.

AFAIK we copied Zephyr's .clang-format file a long time ago and have
since been maintaining a copy of it.

It's not clear to me why we are maintaining a copy instead of syncing
with Zephyr.

I'm assuming this is a mistake.

In this commit we sync with Zephyr.

AFAIK we do not intentionally maintain different formatting from
Zephyr.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>