drivers: wifi: Tie FMAC device management to network interface stats
FMAC device initialization and setup is done as part of init and as
Zephyr doesn't support de-initialization of a network interface the
corresponding FMAC de-initialization and teardown are never invoked.

This means that the RPU is booted once and is always powered on, though
RPU itself does its own power management when low power mode is enabled,
the consumption is low, but even this is unnecessary when an image has
Wi-Fi enabled (hence its initialized) but Wi-Fi is seldom used or unused
completely. Though Wi-Fi itself consumes less power (order of uA) and
QSPI on nRF5340 is in low power state the QSPI consumption is higher in
order mA.

Move the FMAC init/de-init to start/stop of the *first* Wi-Fi interface
so that if all Wi-Fi interfaces are down then RPU can be powered off,
this saves power at the extra cost of time due to re-initialized when
the interface is brought up (patch download, boot etc).

Implements SHEL-1672.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>