bluetooth: hids: Increase max HID report map size from 255 to 512 bytes
In the bluetooth HID service, the size of a report descriptor is held
by a uint8_t, which means that any report descriptors can only have 255
bytes at most. However, the maximum report descriptor size is actually
given by the maximum GATT attribute length, which is 512 bytes.
This commit replaces the uint8_t for a uint16_t and adds an assertion
that enforces the limit set by the Bluetooth specification.

Signed-off-by: Jacob Sánchez <jacobszpz@protonmail.com>