net: azure_iot_hub: Return error if setting backend is not ready
Fix an issue where azure_iot_hub_dps_*_delete() and
azure_iot_hub_dps_reset() APIs would return
zero if -ENOENT was returned by settings_delete().
It was wrongly assumed that the API returned -ENOENT in case the
record didn't exist and hence could not be deleted.
This is not the case. -ENOENT is returned if the settings backend
has not registered with the settings API, which is done in
azure_iot_hub_dps_init().

This patch fixes the issue by propagating -ENOENT to the application.
A test case is added to check that azure_iot_hub_dps_reset() returns
-ENOENT if azure_iot_hub_dps_init() has not been called first.
The settings subsystem is used instead of mock, which in turn requires
the test to only run on native_posix as it has flash capability.

Fixes CIA-996

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>