sysbuild: PM_MCUBOOT_PAD Kconfig setting
Partition manager / signing relies on the size of the MCUboot padding
partition size which is used for the signature header size.

When using dynamic partitions, a default MCUboot pad size of 0x200 is
specified in NCS, and sysbuild defines corresponding setting.

For static partitions, then MCUboot pad size is described in the
partitions file.

Currently, signing of the app happens in the app itself and size is
specified during CMake configure time of the app, where the processing
of partitions file happens after CMake configure has completed for all
images. Until signing, size validation, and dfu packaging has been
reworked for sysbuild the best we therefore can do is to compare the
MCUboot pad size from partitions against the header size used by imgtool
and in case those differs print a warning with information on how to
adust the header size.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>