samples: Fix handling of combo build for Wi-Fi radio test
Wi-Fi radio test combo build is disabled by default and if enabled has
to pull in the peripheral radio test child image to CPUNET for nRF5340.

This is controlled by used of CPUNET configuration, but due to the
mismatch in choosing different Kconfig options RPMSG child image vs.
Peripheral radio test image, the partition manager is enabled because
radio test sample is enabled, but no child image is enabled as CPUNET is
disabled.

Fix this by using the newly add macro that takes care of all
dependencies.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>