lib: at_host: Use zephyr `LOG` for outgoing modem responses
Use zephyr LOG for outgoing modem responses when UART is selected as the
log backend.

This ensures that responses received the modem are scheduled in turn
by the zephyr logging system.

Previously these logs could cut zephyr LOG output because its pushed
directly on UART. This lead to tests failing because the expected
output was corrupted by these modem responses.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>